ALPHA 1.7
The biggest change is the addition of custom shops (working with CustomNPCs mod, now depends on it) and added a forge 1.12.2 .jar with source in order to work with the mod. I remade the menus myself as I figured out that it was not that hard to do. Does not depend on MegaMenus anymore. Lots of progrss are being done, but the plugin is still not ready to be used. Some functions need to be changed in order to be more efficient and to use features the correct way. (Such as data storage / registration)